+++ x/ecocredit/server/migrations.go
@@ -23,8 +23,8 @@ func NewMigrator(keeper Keeper, legacySubspace paramtypes.Subspace) Migrator {
 // Migrate2to3 migrates from version 2 to 3.
 func (m Migrator) Migrate2to3(ctx sdk.Context) error {
 
-	coreStore, basketStore, _ := m.keeper.GetStateStores()
-	if err := v3.MigrateState(ctx, coreStore, basketStore, m.legacySubspace); err != nil {
+	baseStore, basketStore, _ := m.keeper.GetStateStores()
+	if err := v3.MigrateState(ctx, baseStore, basketStore, m.legacySubspace); err != nil {
 		return err
 	}